requirements.txt
Pillow==9.0.1
django-imagekit==4.1.0
